# docs/openapi/baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpecProof Phase 1",
    "description": "Reliable Verification Job API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs/{job_id}/stages": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Job Stages",
        "description": "Pipeline stage timeline for one job (Redis progress stream).\n\nRedis is optional telemetry: when it is down the endpoint still\nanswers 200 with \"degraded: true\" and an empty stage list.",
        "operationId": "job_stages_api_v1_jobs__job_id__stages_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Job Stages Api V1 Jobs  Job Id  Stages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/matrix": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Job Matrix",
        "description": "Requirement-to-evidence matrix rows for one job (contracts table).\n\nThe contracts table stores per-contract results (PASS/FAIL/UNVERIFIED)\nwith requirement text and evidence refs; summary counts are returned\nalongside. Empty rows are honest, not fabricated.",
        "operationId": "job_matrix_api_v1_jobs__job_id__matrix_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Job Matrix Api V1 Jobs  Job Id  Matrix Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/findings": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Job Findings",
        "description": "Findings for one job (summary JSON merged with the findings table).\n\nThe pipeline summary carries descriptions; the findings table adds\ncapsule paths and impact paths. Both may be empty - an empty list is\nan honest answer, 404 is reserved for unknown jobs.",
        "operationId": "job_findings_api_v1_jobs__job_id__findings_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Job Findings Api V1 Jobs  Job Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/certificate": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Job Certificate",
        "description": "Merge Certificate / Rejection Notice for a job, when persisted.\n\nCertificates are written by the CLI verify flow (reports dir). The\nAPI never reconstructs one from scratch: no artifact means 404.",
        "operationId": "job_certificate_api_v1_jobs__job_id__certificate_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Job Certificate Api V1 Jobs  Job Id  Certificate Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/capsule": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "Job Capsule",
        "description": "Download a Bug Capsule zip for a job (404 when none exists).",
        "operationId": "job_capsule_api_v1_jobs__job_id__capsule_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "List Contracts",
        "description": "Contract registry rows with approval-status filtering.",
        "operationId": "list_contracts_api_v1_contracts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "repo_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repo Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Contracts Api V1 Contracts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}